project/requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy>=1.4
psycopg2-binary
alembic
pydantic
python-dotenv
python-multipart
asyncpg
redis    
celery  
python-jose[cryptography] 
requests
pytest
